Do not assume a consul backend
While we currently only define a consul discovery backend, this
removes the assumption that our discovery backend is consul. The
general idea here is that new discovery backends should implement
the DiscoveryService interface, and then config.go is extended
to look for the configuration for that backend in a specific
json field named after the backend ('etcd', 'zookeeper', etc).