Fix failing TestMaintainanceSignal
Trying to test the receiving of signals is subject to race conditions
that cause different test failures in different environments but that
aren't going to cause problems in realistic production environments
where signals we're interested in are either idempotent (SIGTERM) or
can be expected to be received one-at-a-time (SIGUSR1). Integration
tests verify our behavior is correct.

This commit factors out the tests that cover the signal handling code
so that we can test their logic without dealing with coercing the
async signals into a synchronous test model.