Set up global context with accessor and mutex for config reload
Passing around config as a context to functions would be the ideomatic way.
But we need to support configuration reload from signals and have that reload
effect function calls in the main goroutine. Wherever possible we should be
accessing via `getConfig` at the "top" of a goroutine and then use the config
as context for a function after that.