Enable docker-compose to scale app.
The previous docker-compose-local.yml specified the host port for the app service.  As a result, running `docker-compose scale app=3` would fail because the second and third containers would attempt to bind to the same port on the docker host.  By removing the host portion of the port configuration, docker-compose chooses random port numbers for the host when scaling the app.
https://docs.docker.com/compose/compose-file/#ports