Pin GOMAXPROCS=1
The GOMAXPROCS runtime variable is set by default to be equal to the number
of visible CPUs. The runtime uses M:N threading and so we can expect multiple
threads. Containerbuddy has minimal performance requirements, so potentially
creating a thread for each core will increase thread-switching for the main
application that we care about. This might particularly be an issue in
environments like Triton where `runtime.NumCPU` will see all the cores on
the underlying CN and not the number that are likely to be available for
scheduling.

References:

https://golang.org/pkg/runtime/#GOMAXPROCS
https://docs.google.com/document/d/1At2Ls5_fhJQ59kDK2DFVhFu3g5mATSXqqV5QrxinasI/edit