Split maintenance mode lock from signal handler lock.
If our polling functions are waiting on the `inMaintenanceMode` lock, then
they can't select on their `quit` channel. This leads to a deadlock when a
signal handler obtains a r/w lock on the same mutex and then tries to send
on the `quit` channels (ex. `reloadConfig`). By splitting the `maintModeLock`
from the `signalLock`, we can avoid this scenario.

ref https://github.com/joyent/containerbuddy/issues/73.