Swap out log.Fatal in executeAndWait for return code
Under normal circumstances we should only reach the replaced line if the
path to the user-defined executable is misconfigured. But if there's some
other problem with the underlying OS like a disk error then it's possible
to reach it, so we'll rely on the lack of heartbeat to tell us something
went wrong rather than intentionally crash.