Deprecate onStart in lieu of preStart
To improve the clarity of configuration, we're changing the `onStart` option to
`preStart`.

For purposes of backwards compatibility the deprecation behavior is as follows.
If `onStart` is provided without `preStart`, warn the user and alias the
`onStart` to a `preStart`. If both `onStart` and `preStart` are provided, warn
the user and use only the `preStart`.