Catch panic from prometheus.MustRegister
MustRegister panics rather than returning an error (thanks!) if we
register an invalid collector, even just an invalid name. This would give
end-users a big ugly golang stack trace that they have to dig the error
message out of. So we'll catch the panic so we can return an error.