Testing for metrics recording and retrieval
Unit tests of recording metrics for each of the four collector types and
fetching them via the web API. Cleaned up some of the existing tests so that
metrics names are more directly tied-back to test cases.

Fixed a bug in `Sensor.record()` where the type switch doesn't work because
the prometheus collector structs are actually interfaces. So we end up having
to do some ugly string checking.