Back-out reload of telemetry server
The native golang HTTP server cannot be gracefully reloaded. The telemetry
server originally punted on reloading because this meant the only thing we
couldn't change was the telemetry port. Although we made a go at it, we can't
make the reload work so this commit backs out the change (but still uses the
refactored code from PR #138) until we can find a better HTTP server lib to
use.