Remove signal.Reset from config reload
Resetting the signal handlers creates a race that we are having
trouble resolving with creating deadlocks. Fortunately it turns
out that it's unnecessary. When the config is reloaded, we keep
the existing App instance and replace its values. This means we
can safely keep the signal-handling channel around and just reuse
it.