Create generic RunAndWait and RunWithTimeout for exec
This lifts most of the existing logic from the `tasks` package so it can
be reused to give us a single abstraction for running external commands
for user-defined hooks for both blocking actions like the main app and
co-processes (via `RunAndWait`), and non-blocking actions that should
timeout (via `RunWithTimeout`).

This commit includes porting the `services` package to use it.