Update tasks package to use Command wrapper
Most of the internals of `tasks` was moved into `commands`. This commit
also makes the TimeoutDuration of Command public so as to make it
accessible to tests in `tasks`, but in the final cleanup commit of this
effort I'd like to try to avoid that.

Integration tests for tasks have been adjusted for the more consistent
log format.