Refactoring to allow for pollable timeout (#184)
* Move utils exec/run handlers into new commands package

* Create generic RunAndWait and RunWithTimeout for exec

This lifts most of the existing logic from the `tasks` package so it can
be reused to give us a single abstraction for running external commands
for user-defined hooks for both blocking actions like the main app and
co-processes (via `RunAndWait`), and non-blocking actions that should
timeout (via `RunWithTimeout`).

This commit includes porting the `services` package to use it.

* Update backends package to use Command wrapper

* Update coprocess package to use Command wrapper

Added log capture option to RunAndWait, made kill() public.

* Update telemetry package to use Command wrapper

Added RunAndWaitForOutput method to commands.Command and factored out
the setup of the exec.Cmd object (mostly just attaching loggers).

* Update tasks package to use Command wrapper

Most of the internals of `tasks` was moved into `commands`. This commit
also makes the TimeoutDuration of Command public so as to make it
accessible to tests in `tasks`, but in the final cleanup commit of this
effort I'd like to try to avoid that.

Integration tests for tasks have been adjusted for the more consistent
log format.

* Update main and preStart/preStop/postStop to use Command wrapper

* Removed now-unused RunWithFields, ExecuteAndWait

Made sure we had decent test coverage of the new Command methods.

* Move Command.Run* methods into package-level functions

It's too easy to miss a nil check when these methods are called. By
moving these into package-level functions we can centralize the `nil`
check and avoid errors. Callers can choose to ignore the error if its
safe to simply not run the method.