Attach stdout/stderr in RunAndWait (#209)
For https://github.com/joyent/containerpilot/issues/207. We need to
attach stdout/stderr for a command whenever we don't have logging fields
set so that we capture the output of our main application. But we can't
do this in `setUpCmd` because it breaks the case of `RunAndWaitForOutput`
so we've added it to `RunAndWait` directly.