Fix exit codes not being respected by RunWithTimeout
We're assuming that `Process.Wait()` returns an error if the process exits
with an error code. `Cmd.Wait()` returns with an error value on a non-zero
exit code but `Process.Wait()` returns an error on OS-specific details (like
if the process isn't a child process) and instead we should be checking the
value of the `ProcessState.Success()` method.

Added two tests that ensure we don't miss this again.