Send quit to timeout when RunWithTimeout command exits non-zero
When the command executed with RunWithTimeout failed and exited with
a non-zero exit code, we returned early and didn't send on the timeout's
quit channel. This defers the send on that channel right after we set up
the timeout goroutine so that we don't miss any code paths where it needs
to be sent.