Retry heartbeat after initial service registration (#247)
Sending a heartbeat to a not-yet registered service fails on both the
`discovery/consul` and `discovery/etcd` backends. Previously we'd catch
the error, register the service, and return so that we'd send a passing
heartbeat on the next health check poll. This change retries the
heartbeat on the newly registered service so that we become healthy more
quickly.