Fix null-pointer exception in error handling of RunAndWait
`Cmd.Process.Wait` does not guarantee that the `ProcessState` struct it returns
is non-nil if the error it returns is non-nil, because we might have an invalid
PID or a `waitpid` syscall error (ECHILD or EINTR).

We'll check ProcessState struct first b/c it might have the process error code
and this is going to be the common case, but if not then we'll bubble-up the
error. Expanded test coverage for invalid and pruned an unused error handling
branch.