Make unit tests portable across platforms
Many of the unit tests make assumptions about the network interface
names (that they'll match those of a our Docker container). Almost
none of these tests really need to care, so we've replaced them with
the more general "inet" in most places and added a guess for the
localhost name in the one location where we do care.

This commit demonstrates we should decouple the interface name parsing
from the general configuration setup.