Use Hashicorp testutil for testing vs Consul API where required
The behaviors we're testing in the consul package are very thin wrappers
around the upstream Consul API library, but that library doesn't have a
interfaces (in the golang sense) that we can mock. Using httptest with a
test server could give us the HTTP server but not the stateful behaviors
without having to implement a lot of Consul's own behavior and we'd have
to build code generation in order to validate it.

Here we'll use Hashi's own test infrastructure around Consul to run the
Consul binary and tear it down after tests. This lets us run locally
without Docker and makes it clear where the boundary of tests that
depend on the server lies. In future work we can try to minimize this
test region and try to eliminate the requirement or move these pieces
into integration tests.