Introduce control socket and package into configuration parser
This commit introduces a new configuration block within CP3's configuration file
called "control". The goal was to provide a way to override the location of the
Unix socket file within a container. The control plane's HTTP server/endpoints
are served through this socket file. This configuration block is optional and
provides defaults when not overridden by the user. The block can be extended in
the future for things like TLS or HTTP timeout. Finally, this functionality
mirrors the implementation of other CP3 refactored configuration processes found
in other CP packages.

The default location for the socket file is "/var/run/containerpilot.socket".