Introduce and serve HTTPServer as app.ControlSocket
This commit introduces the first pass of an HTTP control plane within the
ContainerPilot process. This is implemented using the configuration additions of
the previous commit. The control socket itself uses a Unix domain socket which
defaults to "/var/run/containerpilot.socket". A goroutine serves HTTP endpoints
tied within the `control/control.go` implementation itself.

Currently, this process only has a single endpoint, GET "/env", which was added
in order to test the functionality and configuration of the control socket. The
goal is to come up with a standard/interfaces for more possible Mariposa HTTP
endpoints.