Initial implementation of HTTP control plane (#300)
* Introduce control socket and package into configuration parser

This commit introduces a new configuration block within CP3's configuration file
called "control". The goal was to provide a way to override the location of the
Unix socket file within a container. The control plane's HTTP server/endpoints
are served through this socket file. This configuration block is optional and
provides defaults when not overridden by the user. The block can be extended in
the future for things like TLS or HTTP timeout. Finally, this functionality
mirrors the implementation of other CP3 refactored configuration processes found
in other CP packages.

The default location for the socket file is "/var/run/containerpilot.socket".

* Introduce and serve HTTPServer as app.ControlSocket

This commit introduces the first pass of an HTTP control plane within the
ContainerPilot process. This is implemented using the configuration additions of
the previous commit. The control socket itself uses a Unix domain socket which
defaults to "/var/run/containerpilot.socket". A goroutine serves HTTP endpoints
tied within the `control/control.go` implementation itself.

Currently, this process only has a single endpoint, GET "/env", which was added
in order to test the functionality and configuration of the control socket. The
goal is to come up with a standard/interfaces for more possible Mariposa HTTP
endpoints.

* Remove control server from app reload

* Address lint issue with constants and comments

* Mount test env endpoint at '/v3'

* Fix default socket test under control config

* Lower case some log messages within control package

* Bump up timeout within TestServiceRunStartupTimeout test

* Mirror testify signature for assert.True and assert.False

* Remove unneeded control config file validation

* Revert "Bump up timeout within TestServiceRunStartupTimeout test"

This reverts commit e68bb400a14e44dec51261fe3ad3e3260df5528d.