Implement new signal based events
This PR implements the process outlined in the tracking issue. I've also
included the ability for running jobs to create an environment variable
including their PID (CONTAINERPILOT_<name>_PID) as well as both the supervisor
and normal worker process signal handlers. So far, the functionality is working
out great and you can fire jobs based on a container receiving a specific signal
(or CP process itself). This is more of a scheduler based feature for CP, acts
as a refill for missing CP2 functionality, and not something particularly useful
for the wider Triton audience.

* Parse signal events in a job's when configuration
* Set the current PID for a job's running command in an env var
* Publish signal events onto the event bus
* Update test_envvars to contain our new PID var
* Add test_sighup integration test
* Generate stringer code for new Signal eventcode
* Supervisor passes signal events through worker process

Fixes: #513