Fix a goroutine leak in the signal handler code path
The root cause of this issue was regarding our change to shutting down via
context. While implementating that change set there was a goroutine leak
introduced that did not provide a clean way for shutting down and reloading the
signal handler. This caused signal handling to be duplicated across reloads.

Fixes: #521