testing: Don't cause NPEs by ignoring errors when configuring Consul
While upgrading Consul to version 1.0.0 I found that our tests caused a null
pointer exception if there's an error configuring a Consul process to test
against. This was due to the fact that we were ignoring errors when configuring
Consul and continuing forward as if Consul was functional. This commit properly
causes a fatal return from the test and includes the actual error returned by
Consul's own internal testing framework (which we use for bootstrapping our
own testing).