Upgrade ContainerPilot to Go 1.9.x and Consul 1.0.0 (#527)
* Update dependencies and peg Consul at v1.0.0
* tools: Switch builds and tooling to use Go 1.9.x and Consul 1.0.0
* testing: Don't cause NPEs by ignoring errors when configuring Consul
* testing: Custom TestServer under `discovery` for exec'ing `consul`
* discovery: Deprecate usage of Consul's internal `testutil` package

This commit upgrades ContainerPilot to be built with Go 1.9 and tested against 
Consul 1.0.0.

While upgrading Consul to version 1.0.0 I found that our tests caused a null
pointer exception if there's an error configuring a Consul process to test
against. This was due to the fact that we were ignoring errors when configuring
Consul and continuing forward as if Consul was functional. This commit properly
causes a fatal return from the test and includes the actual error returned by
Consul's own internal testing framework (which we use for bootstrapping our
own testing).

Finally, this commit removes the use case within the discovery package's tests that
depended on an internal Consul test package, `testutil`. We replace this with
our own TestServer object which is responsible for executing a locally installed
`consul` binary. Consul is installed by our Makefile target `tools` for local
development use.

Fixes: #528