config: Adjust loop function to allow ranges by strings/env vars
This commit adjusts the implementation of the template config function called
`loop`. We make adjustments to allow for the loop function to take a string or
environment variable as input. We use a simple switch statement with a type
assertion and detect either a string or not. By default we continue using `int`
types as normal. Also tests have been adjusted to demonstrate the working
functionality.

Fixes #554